+++ tools/RAiDER/cli/raider.py
@@ -98,6 +98,9 @@ def read_template_file(fname):
         if key == 'download_only':
             template[key] = bool(value)
 
+        if not key in DEFAULT_DICT.keys() and not key.endswith('group'):
+            raise KeyError(f'Invalid key in yaml file: {key}')
+
     # Have to guarantee that certain variables exist prior to looking at heights
     for key, value in params.items():
         if key == 'height_group':
@@ -326,7 +329,7 @@ def calcDelays(iargs=None):
             out_filename = w
 
         if hydro_delay is None:
-            # means that a dataset was returned
+            # means that a dataset was returned (cubes and station files)
             ds = wet_delay
             ext = os.path.splitext(out_filename)[1]
             out_filename = out_filename.replace('wet', 'tropo')